Issue#250: Code refactored to move provider and resolver to separate classes.
