+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/JerseyConfig.java
@@ -1,21 +1,12 @@
 package com.softserveinc.ita.homeproject.application.config;
 
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.stream.Collectors;
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.ext.ContextResolver;
 import javax.ws.rs.ext.Provider;
 
 import org.glassfish.jersey.server.ResourceConfig;
-import org.glassfish.jersey.server.validation.ValidationConfig;
 import org.glassfish.jersey.servlet.ServletProperties;
-import org.hibernate.validator.internal.engine.DefaultParameterNameProvider;
 import org.springframework.context.annotation.ClassPathScanningCandidateComponentProvider;
 import org.springframework.core.type.filter.AnnotationTypeFilter;
 import org.springframework.stereotype.Component;
@@ -49,52 +40,12 @@ public class JerseyConfig extends ResourceConfig {
         scanner.addIncludeFilter(new AnnotationTypeFilter(Provider.class));
         // register endpoints
         this.registerPackageClasses("com.softserveinc.ita.homeproject.application.api", scanner);
-        //  register exception mappers
+        // register exception mappers
         this.registerPackageClasses("com.softserveinc.ita.homeproject.application.exception.mapper", scanner);
         // register providers
         this.registerPackageClasses("com.softserveinc.ita.homeproject.application.provider", scanner);
-        this.register(ValidationConfigurationContextResolver.class);
-    }
-
-    public static class ValidationConfigurationContextResolver implements ContextResolver<ValidationConfig> {
-
-        @Override
-        public ValidationConfig getContext(final Class<?> type) {
-            final ValidationConfig config = new ValidationConfig();
-            config.parameterNameProvider(new RestAnnotationParameterNameProvider());
-            return config;
-        }
-
-        static class RestAnnotationParameterNameProvider extends DefaultParameterNameProvider {
-
-            @Override
-            public List<String> getParameterNames(Method method) {
-                Annotation[][] annotationsByParam = method.getParameterAnnotations();
-                List<String> names = new ArrayList<>(annotationsByParam.length);
-                for (Annotation[] annotations : annotationsByParam) {
-                    String name = getParamName(annotations);
-                    if (name == null) {
-                        name = "arg" + (names.size() + 1);
-                    }
-
-                    names.add(name);
-                }
-
-                return names;
-            }
-
-            private static String getParamName(Annotation[] annotations) {
-                for (Annotation annotation : annotations) {
-                    if (annotation.annotationType() == QueryParam.class) {
-                        return QueryParam.class.cast(annotation).value();
-                    } else if (annotation.annotationType() == PathParam.class) {
-                        return PathParam.class.cast(annotation).value();
-                    }
-                }
-
-                return null;
-            }
-        }
+        // register resolvers
+        this.registerPackageClasses("com.softserveinc.ita.homeproject.application.resolver", scanner);
     }
 
     private void registerPackageClasses(String packageName, ClassPathScanningCandidateComponentProvider scanner) {
